// src/modules/seller-platform/seller-account/enums.ts
export enum SellerType {
  INDIVIDUAL = 'INDIVIDUAL',
  BROKER = 'BROKER',
  AGENCY = 'AGENCY',
}

export enum SellerStatus {
  ACTIVE = 'ACTIVE',
  SUSPENDED = 'SUSPENDED',
}

export enum SellerVerificationStatus {
  UNVERIFIED = 'UNVERIFIED',
  PENDING = 'PENDING',
  VERIFIED = 'VERIFIED',
  REJECTED = 'REJECTED',
}